feat(durham): add `durham-terrain-shaders` with embedded Durham terrain material
Introduce `maybraid/durham/shaders` as a small Bevy crate that embeds the world-space palette-noise WGSL via `embedded_asset!` and registers `DurhamTerrainShader` through `DurhamTerrainShaderPlugin`, so demos and games do not duplicate asset paths. Wire `durham-terrain` as a thin re-export and switch the naturescapes demo to depend on it instead of local `world_space_experiment` sources. This supports [#178](https://github.com/ramate-io/maybraid/issues/178) / [RFC-170 4.7](rfc/rfc-000-000-170-terrain-detail/README.md#47-world-space-ground-color-noise).

## Agent Dialogue

- **Developer:** Reported missing AGENTS.md commit workflow; attached Bevy asset error `Path not found: durham_terrain_shadersdurham_terrain_shader.wgsl` (missing `/` between crate key and file in the embedded URL).
- **Agent:** Refreshed branch `l-monninger/178-world-space-ground-shader` and `HEAD` `b3adf798b629d6e7f12e8d59a99f476e61e26e9c`, wrote this commit message file. Staged `fragment_shader` path already uses `concat!(…, env!("CARGO_CRATE_NAME"), "/", "durham_terrain_shader.wgsl")`; a rebuild should load `embedded://durham_terrain_shaders/durham_terrain_shader.wgsl`. If the error persists, confirm no stale binary and that `embedded_asset!` uses the same relative path as the on-disk `src/durham_terrain_shader.wgsl`.
