+++ demos/naturescapes/src/lib.rs
@@ -5,8 +5,8 @@ pub mod water;
 
 // use bevy::diagnostic::{FrameTimeDiagnosticsPlugin, LogDiagnosticsPlugin};
 use bevy::prelude::*;
+use durham_terrain::shaders::{DurhamTerrainShader, DurhamTerrainShaderPlugin};
 use engine::shaders::{leaf_material::LeafMaterial, outline::EdgeMaterial};
-use shaders::world_space_experiment::WorldSpaceExperimentMaterial;
 use terrain::TerrainSdf;
 
 mod camera;
@@ -22,18 +22,18 @@ impl Plugin for NatureScapesPlugin {
 	fn build(&self, app: &mut App) {
 		// Register material plugins
 		app.add_plugins(bevy::pbr::MaterialPlugin::<EdgeMaterial>::default());
-		app.add_plugins(bevy::pbr::MaterialPlugin::<WorldSpaceExperimentMaterial>::default());
+		app.add_plugins(DurhamTerrainShaderPlugin);
 		app.add_plugins(bevy::pbr::MaterialPlugin::<LeafMaterial>::default());
 		// app.add_plugins(FrameTimeDiagnosticsPlugin::default());
 		// app.add_plugins(LogDiagnosticsPlugin::default());
 		app.add_plugins(water::WaterPlaygroundPlugin);
 		app.add_plugins(terrain::TerrainPlaygroundPlugin {
-			material: WorldSpaceExperimentMaterial::default(),
-			rock_detail_material: WorldSpaceExperimentMaterial::default()
+			material: DurhamTerrainShader::default(),
+			rock_detail_material: DurhamTerrainShader::default()
 				.with_base_color(Vec4::new(0.2, 0.1, 0.1, 1.0)), // reddish brown base color
-			second_rock_detail_material: WorldSpaceExperimentMaterial::default()
+			second_rock_detail_material: DurhamTerrainShader::default()
 				.with_base_color(Vec4::new(0.5, 0.5, 0.5, 1.0)), // gray base color
-			tuft_detail_material: WorldSpaceExperimentMaterial::default()
+			tuft_detail_material: DurhamTerrainShader::default()
 				.with_base_color(Vec4::new(0.2, 0.8, 0.3, 1.0)), // green base color
 		});
 		app.add_plugins(vegetation::VegetationPlaygroundPlugin::<
