+++ demos/naturescapes/src/shaders.rs
@@ -1,3 +1,2 @@
 pub mod refraction_water;
 pub mod water_material;
-pub mod world_space_experiment;
