+++ maybraid/durham/lib/src/lib.rs
@@ -1,0 +1,1 @@
+pub use durham_terrain_shaders as shaders;
