+++ maybraid/durham/shaders/src/lib.rs
@@ -1,0 +1,5 @@
+//! Durham terrain shaders: reusable Bevy materials and embedded WGSL for Maybraid terrain work.
+
+mod durham_terrain_shader;
+
+pub use durham_terrain_shader::{DurhamTerrainShader, DurhamTerrainShaderPlugin};
